muse_ai/backend/requirements.txt
fastapi
uvicorn
torch
music21
google-generativeai
pydantic
